edited smoothing to force integer for window_len

+++ general/smoothing.py
@@ -7,14 +7,14 @@ def gaussian04(N):
 
 def smooth(x,window_len,window):
     """smooth the data using a window with requested size.
-    
+
     This method is based on the convolution of a scaled window with the signal.
-    The signal is prepared by introducing reflected copies of the signal 
+    The signal is prepared by introducing reflected copies of the signal
     (with the window size) in both ends so that transient parts are minimized
     in the begining and end part of the output signal.
-    
+
     input:
-        x: the input signal 
+        x: the input signal
         window_len: the dimension of the smoothing window; should be an odd integer
         window: the type of window from 'flat', 'hanning', 'hamming', 'bartlett', 'blackman'
             flat window will produce a moving average smoothing.
@@ -27,12 +27,12 @@ def smooth(x,window_len,window):
     t=linspace(-2,2,num=1000)
     x=sin(t)+randn(len(t))*0.1
     y=smooth(x)
-    
-    see also: 
-    
+
+    see also:
+
     numpy.hanning, numpy.hamming, numpy.bartlett, numpy.blackman, numpy.convolve
     scipy.signal.lfilter
- 
+
     TODO: the window parameter could be the window itself if an array instead of a string
     NOTE: length(output) != length(input), to correct this: return y[(window_len/2-1):-(window_len/2)] instead of just y.
 
@@ -66,7 +66,6 @@ def smooth(x,window_len,window):
     y=np.convolve(w/w.sum(),s,mode='valid')
 
     if window_len%2 == 0:
-      return y[(window_len/2-1):-(window_len/2)] #NOTE
+      return y[int(window_len/2-1):-int(window_len/2)] #NOTE
     else:
-      return y[(window_len/2):-(window_len/2)]
-
+      return y[int(window_len/2):-int(window_len/2)]
